Split up DiscordPlugin.cs into multiple files before it gets out of hand
