+++ MCGalaxy/MCGalaxy_.csproj
@@ -583,7 +583,9 @@
     <Compile Include="Blocks\Physics\TntPhysics.cs" />
     <Compile Include="Blocks\Physics\ZombiePhysics.cs" />
     <Compile Include="Levels\Zone.cs" />
-    <Compile Include="Modules\DiscordPlugin.cs" />
+    <Compile Include="Modules\Discord\DiscordBot.cs" />
+    <Compile Include="Modules\Discord\DiscordPlugin.cs" />
+    <Compile Include="Modules\Discord\DiscordWebsocket.cs" />
     <Compile Include="Network\BaseWebSocket.cs" />
     <Compile Include="Network\Heartbeat\ClassiCube.cs" />
     <Compile Include="Network\Heartbeat\Heartbeat.cs" />
@@ -720,6 +722,7 @@
     <Folder Include="Games\RoundsGame" />
     <Folder Include="Generator\LibNoise" />
     <Folder Include="Generator\Realistic" />
+    <Folder Include="Modules\Discord" />
     <Folder Include="Network\Heartbeat" />
     <Folder Include="Network\IRCPlugin" />
     <Folder Include="Network\Utils" />
